refactor: add interface folder and constants
