+++ src/types/interface.ts
@@ -1,0 +1,7 @@
+export interface SearchFormProps {
+  onSearch: (searchTerm: string) => void;
+  searchTerm: string;
+}
+export interface SearchFormState {
+  searchTerm: string;
+}
